requirements.txt
langchain>=0.3.0
langchain-openai>=0.3.0
langgraph>=0.2.0
deepagents>=0.1.0
python-dotenv>=1.0.0
